backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.2
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
openai>=1.55.3
httpx<0.28.0